// src/app/shared/models/table.model.ts
export interface TableColumn<T = any> {
  key: keyof T ;
  label: string;
  sortable?: boolean;
  align?: 'left' | 'center' | 'right';
  format?: (value: any, row: T) => string;
}
export type SortDirection = 'asc' | 'desc';

export interface Page<T> {
  content: T[];
  totalElements: number;
  totalPages: number;
  size: number;
  number: number;
  first?: boolean;
  last?: boolean;
}